Set AWS region in environment variable. Calls to S3 are automatically redirected to the correct region, but this saves that extra round-trip if the bucket is not in us-east-1.
Compress backups. The etcd backups compress really well (for a fresh
cluster, 2 MB compresses down to about 80 kB). This will save money on
S3 costs.

Backup all of the kubernetes pki keys. This is needed to properly
restore the cluster. See
https://github.com/cablespaghetti/kubeadm-aws/issues/13.

Compressing the pki data retains the file permissions, so we don't have
to update them when decompressing.

--no-paginate is used to always list all of the objects. By default,
only 1000 objects are returned, so if you changed the cron to take more
frequent backups, then the script would eventually restore an old backup
instead of the most recent backup.

Move instance id to a metadata attribute.